refactor: split data mapping form components into other files
