+++ src/components/config-tabs/numerators/constants.js
@@ -1,0 +1,7 @@
+// Data item types
+export const DATA_ELEMENT = 'dataElement'
+export const INDICATOR = 'indicator'
+
+// Data element types
+export const TOTALS = 'totals'
+export const DETAILS = 'details'
